requirements.txt
ortools==9.11.4210
fastapi==0.115.6
uvicorn==0.34.0
sqlmodel==0.0.22
pydantic_settings==2.7.1
passlib==1.7.4
pyjwt==2.10.1